requirements.txt
azure-functions==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
pdf2image==1.16.0
Pillow==9.0.0
requests==2.27.1
urllib3==1.26.8
